+++ NEWS.md
@@ -1,4 +1,5 @@
-- Soft deprecated `is.term()` or `is_term()`.
+- Soft deprecated `is.incomplete_terms()` for `is_incomplete_terms()`.
+- Soft deprecated `is.term()` for `is_term()`.
 - Soft deprecated 'class' value of `validate` argument of `chk_term()` and `vld_term()` and set to be 'complete'.
 - Added `scalar_term()` to test if each term element is scalar.
 - Added `pars_terms()`.